Replace the sampe script
